JAVA - Quick Quiz + Networking + Multiple Clients
The purpose of quick quiz is for the main client to send questions to multiple sub clients. This happens by connecting to a server. Once the main client is connected the user can send questions to the server. The server will then distribute the questions to the sub clients. Once the sub clients select an answer and submit the question the server will return the answer to the main client where it is displayed in a doubly linked list, a binary tree, and a hashmap. The hashed data is then exported to file.