ASP.NET Core Web App - Quick Quiz
Project: ASP.NET Core - Quick Quiz
Version: 1.0

To use this program open the appsettings file and ensure the connection
string is correct according to the local SQL connection. Open the application
in visual studio and perform an update-database. Once the database has
been added to SQL and the connection string is correct, it is safe to run.

ADMIN LOGIN: 
Username: Admin
Password: 1234
Only the admin has the privelege to edit and delete.

Run both and Front End and the API because the Front end will be communicating
back and forth with the API but the API aswell because it will be communication
to the Front End and the SQL Database.

Once both clients are open, the main application will be hosted as a web app.
I created this app using bootstrap. C# and html were the main languages
used to create the Model-View-Controller.

The purpose of this application to for the user to be able to quickly
create quizzes. Firstly the user will be prompted to login and then immediately
be taken to a page where they are able to create a quiz, then a set of questions,
and then answers. They also have the ability to view all other quizzes from other users.

The key features of the application are:
- ASP.NET Core
- BootStrap
- C#
- HTML
- MVC
- Token Generation to determine administrator priveleges.
- Interactive 2D Graph to display Topics and number of quizzes
- Drag and drop feature

I created this application during my Software Development Diploma which
I completed in TAFE and am very proud of my work.